refactor: split `/create`
