+++ crate-presentation/src/command/create.rs
@@ -1,26 +1,15 @@
-use crate::{
-    global,
-    module::{Module, ModuleExt},
-    shared::{
-        builder::{SlashCommandBuilder, SlashCommandOptionBuilder},
-        command::{CommandResult, ExecutorArgs, InteractionResponse},
-        discord_embeds,
-        ext::{ChronoExt, CommandExt, CreateEmbedExt, IdExt},
-    },
-};
-use crate_domain::{
-    error::MyError, github::Issue as GHIssue, id::IssueId, redmine::Note, status::AgendaStatus,
-};
-use crate_usecase::model::{AgendaDto, DtoExt, RecordParam};
+use crate::shared::builder::{SlashCommandBuilder, SlashCommandOptionBuilder};
 
-use anyhow::{anyhow, ensure, Context};
-use chrono::{Duration, Local, NaiveDate, NaiveDateTime, NaiveTime};
-use futures::stream::{self, StreamExt};
-use itertools::Itertools;
-use log::{debug, info};
-use serenity::{
-    builder::CreateEmbed, model::interactions::application_command::ApplicationCommandOptionType,
-};
+use serenity::model::interactions::application_command::ApplicationCommandOptionType;
+
+mod issue;
+pub use issue::issue;
+
+mod thread;
+pub use thread::thread;
+
+mod new_record;
+pub use new_record::new_record;
 
 pub fn builder() -> SlashCommandBuilder {
     SlashCommandBuilder::new("create", "アイデア会議に関する様々なものを作成します。")
@@ -129,489 +118,3 @@ pub fn builder() -> SlashCommandBuilder {
         )
         .into()
 }
-
-async fn fetch_agendas(
-    module: &Module,
-    id: IssueId,
-) -> (IssueId, Result<AgendaDto, anyhow::Error>) {
-    (id, module.agenda_usecase().find(id).await)
-}
-
-fn refine_all_related_ideas(
-    ideas: Vec<IssueId>,
-    relations: &Vec<IssueId>,
-) -> anyhow::Result<Vec<IssueId>> {
-    let related = ideas
-        .clone()
-        .into_iter()
-        .filter(|id| relations.contains(id))
-        .collect_vec();
-    let not_related = ideas
-        .iter()
-        .filter(|id| !related.contains(id))
-        .collect_vec();
-    ensure!(
-        not_related.is_empty(),
-        "議事録に関連付けられていないチケットがあります。:{:?}",
-        not_related.iter().map(|id| id.formatted()).collect_vec()
-    );
-
-    Ok(related)
-}
-
-fn refine_all_fetched_idea(
-    ideas: Vec<(IssueId, Result<AgendaDto, anyhow::Error>)>,
-) -> anyhow::Result<Vec<AgendaDto>> {
-    let succeeded = ideas
-        .iter()
-        .filter(|(_, res)| res.is_ok())
-        .map(|(_, res)| res.as_ref().unwrap().to_owned())
-        .collect_vec();
-    let failed = ideas
-        .iter()
-        .filter(|(id, _)| !succeeded.iter().map(|v| v.id).contains(id))
-        .map(|(id, res)| (id, res.as_ref().err().unwrap()))
-        .collect_vec();
-    ensure!(
-        failed.is_empty(),
-        "詳細を取得できない議題があります。:{:?}",
-        failed
-            .iter()
-            .map(|(id, err)| format!("{} {:?}", id.formatted(), err))
-            .collect_vec()
-    );
-
-    Ok(succeeded)
-}
-
-fn refine_all_approved_ideas(ideas: Vec<AgendaDto>) -> anyhow::Result<Vec<AgendaDto>> {
-    let not_approved = ideas
-        .iter()
-        .filter(|v| v.status != AgendaStatus::Approved)
-        .collect_vec();
-    let approved = ideas
-        .iter()
-        .filter(|v| !not_approved.contains(v))
-        .map(|dto| dto.to_owned())
-        .collect_vec();
-    ensure!(
-        not_approved.is_empty(),
-        "承認されていない議題があります。:{:?}",
-        not_approved.iter().map(|v| v.id.formatted()).collect_vec()
-    );
-
-    Ok(approved)
-}
-
-/// 指定された議題を連結した文字列が当該議事録の承認された議題として正しいかどうかを確認する
-///
-/// すべての議題が以下の条件を満たす必要がある
-/// * u16にパースできる
-/// * 議事録に関連付けられている
-/// * ステータスが承認である
-///
-/// ## 引数
-///
-/// * `idea_args` - 議題のチケット番号をスペース区切りでつなげた文字列
-/// * `relations` - 議事録の関連チケットID
-/// * `module` - ユースケースを解決するModule
-async fn fetch_ideas(
-    idea_args: String,
-    relations: &Vec<IssueId>,
-    module: &Module,
-) -> anyhow::Result<Vec<AgendaDto>> {
-    debug!("議題文字列: {:?}", idea_args);
-    let ideas = idea_args
-        .split(' ')
-        .filter_map(|str| str.parse::<u16>().ok())
-        .map(IssueId::new)
-        .collect_vec();
-    ensure!(
-        !ideas.is_empty(),
-        "指定された文字列を議題のリストとして認識できません。"
-    );
-
-    let related = refine_all_related_ideas(ideas, relations)?;
-
-    let fetch_agenda_results: Vec<_> = stream::iter(related)
-        .then(|id| fetch_agendas(&module, id))
-        .collect()
-        .await;
-    let fetched = refine_all_fetched_idea(fetch_agenda_results)?;
-
-    refine_all_approved_ideas(fetched)
-}
-
-async fn create_gh_issue(
-    id: IssueId,
-    issue: GHIssue,
-    module: &Module,
-) -> (IssueId, anyhow::Result<String>) {
-    (id, module.gh_issue_usecase().add(issue).await)
-}
-
-fn refine_gh_issues(
-    issues: Vec<(IssueId, anyhow::Result<String, anyhow::Error>)>,
-) -> anyhow::Result<Vec<(IssueId, String)>> {
-    let succeeded = issues
-        .iter()
-        .filter(|(_, res)| res.is_ok())
-        .map(|(id, res)| (id.to_owned(), res.as_ref().unwrap().to_owned()))
-        .collect_vec();
-    let failed = issues
-        .iter()
-        .filter(|(_, res)| res.is_err())
-        .map(|(id, res)| (id.to_owned(), res.as_ref().err().unwrap()))
-        .collect_vec();
-    ensure!(
-        failed.is_empty(),
-        "GitHubにIssueを起票できなかった議題があります。:{:?}",
-        failed
-    );
-
-    Ok(succeeded)
-}
-
-async fn add_redmine_notes(
-    id: IssueId,
-    note: Note,
-    module: &Module,
-) -> (IssueId, anyhow::Result<()>) {
-    (id, module.record_usecase().add_note(id, note).await)
-}
-
-fn refine_redmine_notes(notes: Vec<(IssueId, anyhow::Result<()>)>) -> anyhow::Result<Vec<IssueId>> {
-    let succeeded = notes
-        .iter()
-        .filter(|(_, res)| res.is_ok())
-        .map(|(id, _)| id.to_owned())
-        .collect_vec();
-    let failed = notes
-        .iter()
-        .filter(|(_, res)| res.is_err())
-        .map(|(id, res)| (id.to_owned(), res.as_ref().err().unwrap()))
-        .collect_vec();
-    ensure!(
-        failed.is_empty(),
-        "Redmineに注記できなかった議題があります。:{:?}",
-        failed
-    );
-
-    Ok(succeeded)
-}
-
-#[allow(clippy::type_complexity)]
-pub async fn issue((map, ctx, interaction): ExecutorArgs) -> CommandResult {
-    let module = global::module::get();
-
-    info!("Start /create issue");
-
-    // 議事録のIDを取得
-    let record_id: u16 = map
-        .get("record_issue_number")
-        .ok_or_else(|| MyError::ArgIsNotFound("record_issue_number".to_string()))?
-        .to_owned()
-        .try_into()?;
-    let record = module
-        .record_usecase()
-        .find(IssueId::new(record_id))
-        .await
-        .with_context(|| format!("議事録の取得中にエラーが発生しました。: #{:?}", record_id))?;
-
-    // Issueを作成するアイデアを取得
-    let ideas: String = map
-        .get("idea_issue_numbers")
-        .ok_or_else(|| MyError::ArgIsNotFound("idea_issue_numbers".to_string()))?
-        .to_owned()
-        .try_into()?;
-    let ideas = fetch_ideas(ideas, &record.relations, &module).await?;
-
-    info!("Create GitHub issues");
-    // GitHubにIssueを作成
-    let gh_issues = ideas
-        .iter()
-        .map(|idea| {
-            let title = format!("Redmine Idea {}", idea.id.formatted());
-            let content = format!(
-                "{}\n[{}]({})にて承認されたアイデア。",
-                idea.url(),
-                record.discussion_title(),
-                record.url()
-            );
-
-            (
-                idea.id,
-                GHIssue::new(
-                    title,
-                    content,
-                    vec!["Tracked: Redmine", "Status/Idea: Accepted✅"]
-                        .into_iter()
-                        .map(|str| str.to_string())
-                        .collect_vec(),
-                ),
-            )
-        })
-        .collect_vec();
-    let github_result: Vec<_> = stream::iter(gh_issues)
-        .then(|(id, issue)| create_gh_issue(id, issue, &module))
-        .collect()
-        .await;
-    let gh_issues = refine_gh_issues(github_result)?;
-
-    info!("Add Redmine notes");
-    // RedmineにGitHubのIssueのURLを注記
-    let redmine_notes = gh_issues
-        .iter()
-        .map(|(id, gh_issue_url)| {
-            (
-                id.clone(),
-                Note::new(
-                    vec![
-                        "GitHubにIssueを作成しました。以下URLより確認できます。",
-                        gh_issue_url,
-                    ]
-                    .into_iter()
-                    .map(|s| s.to_string())
-                    .collect_vec(),
-                ),
-            )
-        })
-        .collect_vec();
-    let redmine_result: Vec<_> = stream::iter(redmine_notes)
-        .then(|(id, note)| add_redmine_notes(id, note, &module))
-        .collect()
-        .await;
-    let redmine_notes = refine_redmine_notes(redmine_result)?;
-
-    let result_embed = CreateEmbed::default()
-        .custom_default(&record.id)
-        .title("GitHubへの起票とRedmineへの注記を行いました")
-        .custom_field(
-            "処理を開始した議題",
-            ideas.into_iter().map(|idea| idea.id.formatted()).join(", "),
-            false,
-        )
-        .custom_field(
-            "GitHubにIssueを作成した議題",
-            gh_issues
-                .into_iter()
-                .map(|(id, _)| id.formatted())
-                .join(", "),
-            false,
-        )
-        .custom_field(
-            "Redmineに注記をした議題",
-            redmine_notes
-                .into_iter()
-                .map(|id| id.formatted())
-                .join(", "),
-            false,
-        )
-        .success_color()
-        .to_owned();
-
-    interaction
-        .send(&ctx.http, InteractionResponse::Embed(result_embed))
-        .await
-        .map(|_| ())
-}
-
-pub async fn thread((map, ctx, interaction): ExecutorArgs) -> CommandResult {
-    let module = global::module::get();
-
-    info!("Create threads");
-
-    // 議事録のIDを取得
-    let record_id: u16 = map
-        .get("record_issue_number")
-        .ok_or_else(|| MyError::ArgIsNotFound("record_issue_number".to_string()))?
-        .to_owned()
-        .try_into()?;
-    let record = module
-        .record_usecase()
-        .find(IssueId::new(record_id))
-        .await
-        .with_context(|| format!("議事録の取得中にエラーが発生しました: #{:?}", record_id))?;
-
-    // スレッドを作成するアイデアを取得
-    // ただし、以下をすべて満たす必要がある
-    // * u16にパースできる
-    // * 議事録に関連付けられている
-    // * ステータスが承認である
-    let ideas: String = map
-        .get("idea_issue_numbers")
-        .ok_or_else(|| MyError::ArgIsNotFound("idea_issue_numbers".to_string()))?
-        .to_owned()
-        .try_into()?;
-    let ideas = ideas
-        .split(' ')
-        .filter_map(|str| str.parse::<u16>().ok())
-        .map(IssueId::new)
-        .filter(|id| record.relations.contains(id))
-        .collect_vec();
-    let ideas: Vec<_> = stream::iter(ideas)
-        .then(|id| module.agenda_usecase().find(id))
-        .collect()
-        .await;
-    let ideas = ideas
-        .into_iter()
-        .filter_map(|res| res.ok())
-        .filter(|idea| idea.status == AgendaStatus::Approved)
-        .collect_vec();
-
-    debug!("ideas: {:?}", ideas);
-    ensure!(
-        !ideas.is_empty(),
-        anyhow!("指定された議題は、いずれも存在しないか条件を満たしていません。")
-    );
-
-    let base_msg = interaction
-        .send(
-            &ctx.http,
-            InteractionResponse::Message("スレッドを作成しました。".to_string()),
-        )
-        .await?;
-
-    info!("Start to create threads per each idea");
-    for idea in ideas.iter() {
-        debug!("Create thread of {}", idea.id.formatted());
-
-        if let Ok(th) = interaction
-            .channel_id
-            .create_public_thread(&ctx.http, base_msg.id, |b| {
-                // Threads will be archived in 24 hours automatically
-                b.name(format!(
-                    "{}: {}",
-                    record.discussion_title(),
-                    idea.id.formatted()
-                ))
-                .auto_archive_duration(60 * 24)
-            })
-            .await
-        {
-            let _ = th
-                .send_message(&ctx.http, |b|
-                    b.content(format!(
-                        "このスレッドは、{}にて承認されたアイデアについて個別に議論を行うためのものです。",
-                         record.discussion_title()
-                    )).embed(|e|
-                        discord_embeds::next_agenda_embed(e, &record.id, idea)
-                            .title(format!("このスレッドで議論を行う議題は{}です", idea.id.formatted()))
-                    )
-                )
-                .await;
-        }
-    }
-
-    Ok(())
-}
-
-pub async fn new_record((map, ctx, interaction): ExecutorArgs) -> CommandResult {
-    let module = global::module::get();
-
-    info!("Create new record");
-
-    // 次回の会議の日付・時刻を取得
-    let date = {
-        let arg_strs = vec!["next_date_year", "next_date_month", "next_date_day"];
-        let mut args = Vec::new();
-        for str in arg_strs.into_iter() {
-            let arg: u16 = map
-                .get(str)
-                .cloned()
-                .ok_or_else(|| MyError::ArgIsNotFound(str.to_string()))?
-                .try_into()
-                .unwrap();
-            args.push(arg);
-        }
-
-        NaiveDate::from_ymd(args[0].into(), args[1].into(), args[2].into())
-    };
-    let start_time = {
-        let args: Vec<u16> = vec![("next_time_hour", 21), ("next_time_minute", 0)]
-            .into_iter()
-            .map(|(str, default)| {
-                map.get(str)
-                    .cloned()
-                    .map(|arg| arg.try_into().unwrap())
-                    .unwrap_or(default)
-            })
-            .collect_vec();
-
-        NaiveTime::from_hms(args[0].into(), args[1].into(), 0)
-    };
-    let end_time = start_time + Duration::hours(2);
-
-    // 次回の会議の日付・時刻を文字列にフォーマット
-    let date_str = format!("{}({})", date.format("%Y/%m/%d"), date.weekday_ja());
-    let time_formatter = "%H:%M";
-    let start_time_str = start_time.format(time_formatter);
-    let end_time_str = end_time.format(time_formatter);
-
-    info!("{} {} - {}", date_str, start_time_str, end_time_str);
-    ensure!(
-        Local::now().naive_local() <= NaiveDateTime::new(date, start_time),
-        format!(
-            "現在または現在より未来の日時を指定してください。: {} {}",
-            date_str, start_time_str
-        )
-    );
-
-    // 次回の会議の回数を取得
-    let latest_closed_discussion_number = module
-        .record_usecase()
-        .find_latest_closed()
-        .await
-        .context("No closed record")?
-        .discussion_number()
-        .context("Error while getting latest record number")?;
-    let next_discussion_number = latest_closed_discussion_number + 1;
-
-    debug!(
-        "latest {} -> next {}",
-        latest_closed_discussion_number, next_discussion_number
-    );
-
-    // 議事録のタイトルと説明文を生成
-    let record_title = format!("{}　第{}回アイデア会議", date_str, next_discussion_number);
-    let record_description_date_time =
-        format!("{}\n{}〜{}\n", date_str, start_time_str, end_time_str);
-    let record_description = format!("{}{}", record_description_date_time, RECORD_DESCRIPTIONS);
-
-    // 議事録をRedmine上に作成
-    let new_record_param = RecordParam::new(record_title, record_description, None, Some(date));
-    let new_record = module.record_usecase().add(new_record_param).await?;
-
-    let embed = CreateEmbed::default()
-        .title("議事録を新規作成しました")
-        .custom_field("タイトル", &new_record.title, false)
-        .custom_field("URL", new_record.url(), false)
-        .current_timestamp()
-        .success_color()
-        .to_owned();
-
-    interaction
-        .send(&ctx.http, InteractionResponse::Embed(embed))
-        .await
-        .map(|_| ())
-}
-
-const RECORD_DESCRIPTIONS: &str = r"にアイデア会議を行います。
-
-**開始前**
-* 議題は予めこちらのチケットに注記を付加する形で提唱願います。
-* ただし、運営チーム等が別途必要に応じて募集したアイデアに関しての議論は行いません。
-
-**当日**
-* 会議開始前に司会及び議事録作成係を決めて下さい。
-* 議題順番は必要に応じて入れ替えてOKです。
-* 議事録作成者は案件毎に「簡単な議論の流れ(後述する結論への補足として)」「結論(こういう方向で進みます、保留にします等)」「(結論でタスクが出来た場合)作業する担当者」「(ある場合)期日」などを議事録へ記載して下さい(内容は適宜調節してOKです。)
-* 会議中は議事録作成者の負担を考慮して下さい。話を早く進めすぎると議事録への反映が漏れたり、異なる内容が記録される原因となります。必要に応じて議事録へ記載して欲しい内容を議事録作成者に伝えると良いと思います。
-
-**終了後**
-* 議事録作成係は作成した議事録を本チケットに注記を付加する形で投稿して下さい。
-(会議に参加出来なかった方は後日議事録をご確認下さい)
-
-**その他**
-* 会議の進め方や心得が[[Wiki]]にまとまっていますのでご覧下さい。
-* 議題として消化されたチケットは、運営チームが更新(クローズするなど)をしてください。";
